started separating data logic and display logic
