+++ src/components/TeamMemberTasks/actions.js
@@ -1,0 +1,3 @@
+import { createAction } from "redux-actions";
+
+export const fetchTeamMembersTaskSuccess = createAction("FETCH_TEAM_MEMBERS_TASK_SUCCESS");
