+++ src/components/TeamMemberTasks/selectors.js
@@ -1,0 +1,1 @@
+export const getTeamMemberTasksData = (state) => state.tasks;
